[redhat] configs: enable CONFIG_PKEY for s390x
Message-id: <1507295118-24634-58-git-send-email-brueckner@redhat.com>
Patchwork-id: 189665
O-Subject: [RHEL7.5 PATCH 57/57] [redhat] configs: enable CONFIG_PKEY for s390x
Bugzilla: 1380349
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
Enable and build the pkey kernel module to provide an interface
for protected and secure key handling.

Note that for the -kdump (zfcpdump) kernel variant, this is not
enabled because the pkey configuration depends on the zcrypt
device driver.  The zcrypt device driver is explicitly disabled
for the -kdump variant.

Bugzilla
--------
BZ 1380349
https://bugzilla.redhat.com/show_bug.cgi?id=1380349

Upstream status of the patch
----------------------------
This is RHEL7-specific kernel configuration.  The upstream kernel also
builds pkey as kernel module according to the s390 defconfig.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=14193246

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>