[x86] dumpstack: Remove raw stack dump
Message-id: <20171006175802.27157-1-swood@redhat.com>
Patchwork-id: 5610
O-Subject: [kernel team] [RHEL7.5 BZ 1479560] x86/dumpstack: Remove raw stack dump
Bugzilla: 1479560
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1479560

https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14187776

There is a race condition where an NMI-triggered stack trace hits the
first couple instructions of the syscall entry path, before the kernel
stack pointer has been loaded.  The stack trace itself validates that the
stack pointer is on one of the expected kernel stacks, but the raw dump
of stack contents does not.  Besides the potential for malicious
userspace to get the kernel to read from an arbitrary kernel pointer
(which could be I/O with a side effect, or an information leakage via
dmesg), crashes have been seen on machines with the Supervisor Mode
Access Prevention mechanism due to the unexpected kernel access of a
userspace address.

Testing: On a machine with SMAP (dell-per230-04.khw.lab.eng.bos.redhat.com),
loaded each CPU with a gettimeofday loop, verified with strace that the
vdso wasn't being used (which was a bit of a surprise), and did 'echo l >
/proc/sysrq-trigger' in a loop.  Before this patch it quickly caused an
oops.  After, no oops despite seeing traces in dmesg that show the race.

commit 0ee1dd9f5e7eae4e55f95935b72d4beecb03de9c
Author: Josh Poimboeuf <jpoimboe@redhat.com>
Date:   Tue Oct 25 09:51:13 2016 -0500

    x86/dumpstack: Remove raw stack dump

    For mostly historical reasons, the x86 oops dump shows the raw stack
    values:

      ...
      [registers]
      Stack:
       ffff880079af7350 ffff880079905400 0000000000000000 ffffc900008f3ae0
       ffffffffa0196610 0000000000000001 00010000ffffffff 0000000087654321
       0000000000000002 0000000000000000 0000000000000000 0000000000000000
      Call Trace:
      ...

    This seems to be an artifact from long ago, and probably isn't needed
    anymore.  It generally just adds noise to the dump, and it can be
    actively harmful because it leaks kernel addresses.

    Linus says:

      "The stack dump actually goes back to forever, and it used to be
       useful back in 1992 or so. But it used to be useful mainly because
       stacks were simpler and we didn't have very good call traces anyway. I
       definitely remember having used them - I just do not remember having
       used them in the last ten+ years.

       Of course, it's still true that if you can trigger an oops, you've
       likely already lost the security game, but since the stack dump is so
       useless, let's aim to just remove it and make games like the above
       harder."

    This also removes the related 'kstack=' cmdline option and the
    'kstack_depth_to_print' sysctl.

    Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
    Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
    Cc: Andy Lutomirski <luto@kernel.org>
    Cc: Borislav Petkov <bp@alien8.de>
    Cc: Brian Gerst <brgerst@gmail.com>
    Cc: Denys Vlasenko <dvlasenk@redhat.com>
    Cc: H. Peter Anvin <hpa@zytor.com>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Thomas Gleixner <tglx@linutronix.de>
    Link: http://lkml.kernel.org/r/e83bd50df52d8fe88e94d2566426ae40d813bf8f.1477405374.git.jpoimboe@redhat.com
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

 Conflicts:
	Documentation/x86/x86_64/boot-options.txt
	arch/x86/include/asm/stacktrace.h
	arch/x86/kernel/dumpstack.c
	arch/x86/kernel/dumpstack_32.c
	arch/x86/kernel/dumpstack_64.c

The untrusted pointer dereference does not seem like that much of a risk
in practice, given that it depends on an NMI backtrace being triggered,
but sending to rhkernel-team-list just in case.

Signed-off-by: Scott Wood <swood@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>