[virt] vfio: New external user group/file match
Message-id: <20171006202257.9395.55132.stgit@gimli.home>
Patchwork-id: 189827
O-Subject: [RHEL-7.5 kernel PATCH 04/11] vfio: New external user group/file match
Bugzilla: 1499022
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1499022
Upstream: 5d6dee80a1e94cc284d03e06d930e60e8d3ecf7d

At the point where the kvm-vfio pseudo device wants to release its
vfio group reference, we can't always acquire a new reference to make
that happen.  The group can be in a state where we wouldn't allow a
new reference to be added.  This new helper function allows a caller
to match a file to a group to facilitate this.  Given a file and
group, report if they match.  Thus the caller needs to already have a
group reference to match to the file.  This allows the deletion of a
group without acquiring a new reference.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Reviewed-by: Paolo Bonzini <pbonzini@redhat.com>
Tested-by: Eric Auger <eric.auger@redhat.com>
Cc: stable@vger.kernel.org
Signed-off-by: Rafael Aquini <aquini@redhat.com>