[virt] vfio: Stall vfio_del_group_dev() for container group detach
Message-id: <20171006202417.9395.22449.stgit@gimli.home>
Patchwork-id: 189832
O-Subject: [RHEL-7.5 kernel PATCH 09/11] vfio: Stall vfio_del_group_dev() for container group detach
Bugzilla: 1499022
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1499022
Upstream: 6586b561a91cd80a91c8f107ed0d144feb3eadc2

When the user unbinds the last device of a group from a vfio bus
driver, the devices within that group should be available for other
purposes.  We currently have a race that makes this generally, but
not always true.  The device can be unbound from the vfio bus driver,
but remaining IOMMU context of the group attached to the container
can result in errors as the next driver configures DMA for the device.

Wait for the group to be detached from the IOMMU backend before
allowing the bus driver remove callback to complete.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>