[s390] lib: export udelay_simple for systemtap
Message-id: <1507559525-28712-1-git-send-email-brueckner@redhat.com>
Patchwork-id: 190653
O-Subject: [RHEL7.5 PATCH] [s390] lib: export udelay_simple for systemtap
Bugzilla: 1499014
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
A kernel bug due to a spinlock recursion occured when using systemtap
to set particular kprobes.  The problem is that a kprobe was set on
lock_acquired() when the corresponding probe was hit.  The generated
probe code calls udelay().  The systemtap's runtime assume that udelay()
is an atomic busy-loop.  The udelay() function is used inn systemtap,
for example, when contention is detected over a shared variable ("dummy"),
and it must delay a while.   See "stap -p3" code related to MAXTRYLOCK.

The s390 implementation of __udelay() sets up an (internal) timer and
reenables interrupts (actually just the timer interrupt) and waits for
the interrupt.  This is an optimization to avoid that the cpu is busy
looping while waiting that enough time passes. The problem now is that
the interrupt handler still does call irq_enter()/irq_exit() which then
again leads to the observed deadlock.

A patch which fixes the s390 udelay implementation went upstream:

https://git.kernel.org/cgit/linux/kernel/git/s390/linux.git/commit?id=db7e007fd61ca07e5f3b04cf2168531b074d318d

Note that a backport of this to RHEL7 might break other things in a
subtle way.  Therefore, do the same as on RHEL6 and export the
udelay_simple() function and use that instead.

Bugzilla
--------
BZ 1499014
https://bugzilla.redhat.com/show_bug.cgi?id=1499014

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=14218704

Signed-off-by: Rafael Aquini <aquini@redhat.com>