[scsi] lpfc: Fix driver load issues when MRQ=8
Message-id: <1507670879-17640-57-git-send-email-dkennedy@redhat.com>
Patchwork-id: 191067
O-Subject: [RHEL7.5 e-stor PATCH 056/170] Fix driver load issues when MRQ=8
Bugzilla: 1385844 1461977 1387768
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

The symptom is that the driver will fail to login to the fabric.
The reason is because it is out of iocb resources.

There is a one to one relationship between MRQs
(receive buffers for NVMET-FC) and iocbs and the default number of
IOCBs was not accounting for the number of MRQs that were being created.

This fix aligns the number of MRQ resources with the total resources so
that it can handle fabric events when needed.

Also the initialization of ctxlock to be on FCP commands, NOT LS commands.
And modified log messages so that the log output can be correlated with
the analyzer trace.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <james.smart@broadcom.com>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
(cherry picked from commit 2b7824d00d9cae2e625dcf6b004c5d738a9c8df4)
Signed-off-by: Rafael Aquini <aquini@redhat.com>