[scsi] lpfc: Fix transition nvme-i rport handling to nport only
Message-id: <1507670879-17640-93-git-send-email-dkennedy@redhat.com>
Patchwork-id: 191102
O-Subject: [RHEL7.5 e-stor PATCH 092/170] scsi: lpfc: Fix transition nvme-i rport handling to nport only.
Bugzilla: 1385844 1461977 1387768
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

As the devloss API was implemented in the nvmei driver, an evaluation of
the nvme transport and the lpfc driver showed dual management of the
rports.  This creates a bug possibility when the thread count and SAN
size increases.

The nvmei driver code was based on a very early transport and was not
revisited until the devloss API was introduced.

Remove the listhead in the driver's rport data structure and the
listhead in the driver's lport data structure.  Remove all rport_list
traversal.  Convert the driver to use the nrport (nvme rport) pointer
that is now NULL or nonNULL depending on a devloss action.  Convert
debugfs and nvme_info in sysfs to use the fc_nodes list in the vport.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <james.smart@broadcom.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 80cc004393619a1b3a17aaf4a9e55c5b9f4fc3c1)
Signed-off-by: Rafael Aquini <aquini@redhat.com>