[scsi] lpfc: Fix PRLI retry handling when target rejects it
Message-id: <1507670879-17640-103-git-send-email-dkennedy@redhat.com>
Patchwork-id: 191122
O-Subject: [RHEL7.5 e-stor PATCH 102/170] scsi: lpfc: Fix PRLI retry handling when target rejects it.
Bugzilla: 1385844 1461977 1387768
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

The nvmet driver was rejecting the initiator's PRLI because its reg_rpi
for the PLOGI was still outstanding.  The initiator would resend the
PRLI without delay and get the same answer.  The PRLI retries would
exhaust causing the nvme initiator to set the nvmet ndlp to UNMAPPED.

The driver's lpfc_els_retry handler did not have a policy for an LS_RJT
with explanation CMD_IN_PROGRESS for PRLI or NVME_PRLI.  This caused the
delay to remain at 0 but retry set 1.

Fix: When the ELS response is LS_RJT, TPC and the command was PRLI or
NVME_PRLI, just set the delay to 1000 mS to get a 1 second delay on the
PRLI retry.  This was enough to allow the REG_RPI to complete at the
target.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <james.smart@broadcom.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit dea37e82fa3423a6c8d1325d44e68b6d03892453)
Signed-off-by: Rafael Aquini <aquini@redhat.com>