[sound] alsa: control: queue events within locking of controls_rwsem for ELEM_WRITE operation
Message-id: <1507829954160572-85-patchset-jkysela@redhat.com>
Patchwork-id: 191989
O-Subject: [RHEL 7.5 PATCH v2 0084/1088] ALSA: control: queue events within locking of controls_rwsem for ELEM_WRITE operation
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Any control event is queued by a call of snd_ctl_notify(). This function
adds the event to each queue of opened file data corresponding to ALSA
control character devices. This function acquired two types of lock; a
counting semaphore for a list of the opened file data and a spinlock for
card data opened by the file. Typically, this function is called after
acquiring a counting semaphore for a list of elements in the card data.

In current implementation of a handler for ELEM_WRITE request, the
function is called after releasing the semaphore for a list of elements
in the card data. This release is not necessarily needed.

This commit removes the release to call the function within the critical
section so that later commits are simple.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7b42cfafdcbf73bd58687cbe3157b9ca4a0fd2e5)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>