[sound] alsa: control: queue TLV event for a set of user-defined element
Message-id: <1507829954160572-92-patchset-jkysela@redhat.com>
Patchwork-id: 191992
O-Subject: [RHEL 7.5 PATCH v2 0091/1088] ALSA: control: queue TLV event for a set of user-defined element
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In a design of user-defined element set, applications allow to change TLV
data on the set. This operation doesn't only affects to a target element,
but also to elements in the set.

This commit generates TLV event for all of elements in the set when the TLV
data is changed.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit da4288287b68fe6902629f4e5306aba2a554bc4b)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>