[sound] alsa: of: add helper for getting endpoint node of specific identifiers
Message-id: <1507829954160572-473-patchset-jkysela@redhat.com>
Patchwork-id: 192385
O-Subject: [RHEL 7.5 PATCH v2 0472/1088] ALSA: of: add helper for getting endpoint node of specific identifiers
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When there are multiple ports or multiple endpoints in a port, they have to be
distinguished by the value of reg property. It is common. The drivers can get
the specific endpoint in the specific port via this function. Now the drivers
have to implement this code in themselves or have to force the order of dt nodes
to get the right node.

Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>
Acked-by: Rob Herring <robh+dt@kernel.org>
Signed-off-by: Inki Dae <inki.dae@samsung.com>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 8ccd0d0ca04147e91890c373677f1e741dda2631)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>