[sound] alsa: asoc: intel: skylake: Add support for deferred DSP module bind
Message-id: <1507829954160572-616-patchset-jkysela@redhat.com>
Patchwork-id: 192519
O-Subject: [RHEL 7.5 PATCH v2 0615/1088] ALSA: ASoC: Intel: Skylake: Add support for deferred DSP module bind
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Module at the end of DSP pipeline that needs to be connected to a module
in another pipeline are represented as a PGA(leaf node) and in PGA event
handler these modules are bound/unbounded. Modules other than PGA leaf
can be connected directly or via switch to a module in another pipeline.
Example: reference path.

To support the deferred DSP module bind, following changes are done:
o When the path is enabled, the destination module that needs to be
bound may not be initialized. If the module is not initialized, add
these modules in a deferred bind list.
o When the destination module is initialized, check for these modules
in deferred bind list. If found, bind them.
o When the destination module is deleted, Unbind the modules.
o When the source module is deleted, remove the entry from the deferred
bind list.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Acked-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit b8c722ddd548186db3f62e64f727af4db1cf9517)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>