[sound] alsa: asoc: intel: bxt: Move codec sysclk config to codec_init function
Message-id: <1507829954160572-665-patchset-jkysela@redhat.com>
Patchwork-id: 192535
O-Subject: [RHEL 7.5 PATCH v2 0664/1088] ALSA: ASoC: Intel: bxt: Move codec sysclk config to codec_init function
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The MCLK for DA7219 does not change in this platform, but is
currently being configured everytime as part of the platform_clock
event handler for DAPM. The upshot of this is that we have
unnecessary calls to this function, and it also means that if
a stream hasn't yet been started, DA7219 driver does not have the
correct MCLK rates programmed and so the HP detection feature does
not operate as expected.

This patch rectifies this issue by moving the sysclk call to
codec_init function so it's only called once at initialisation.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Acked-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit fd0f237572ada2b18fe1e8684ae9fa131b577a17)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>