[sound] alsa: asoc: da7219: Make more efficient use of MCLK within driver
Message-id: <1507829954160572-873-patchset-jkysela@redhat.com>
Patchwork-id: 192591
O-Subject: [RHEL 7.5 PATCH v2 0872/1088] ALSA: ASoC: da7219: Make more efficient use of MCLK within driver
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Currently, if the driver has control of MCLK then it remains
enabled as long as the codec is in STANDBY or above. The MCLK is
only really required in STANDBY when a 3-pole jack is inserted
and the HP detect procedure is required to run.

This patch updates the code to enable/disable the MCLK when moving
between the STANDBY and PREPARE bias level, and when a 3-pole jack
is inserted and HP detection is required, thus saving power at all
other times.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 40585391fc88d6d66dc479efccba973426c004ab)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>