[sound] alsa: asoc: rt5514: Unconfuse the rt5514 at probe / resume time
Message-id: <1507829954160572-1034-patchset-jkysela@redhat.com>
Patchwork-id: 192659
O-Subject: [RHEL 7.5 PATCH v2 1033/1088] ALSA: ASoC: rt5514: Unconfuse the rt5514 at probe / resume time
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The rt5514 can get confused and incorrectly detect a start bit if the
SCL/SDA lines happen to both go low and then high again.  This
situation has been seen to happen at reboot time and is also
theoretically possible during suspend/resume if the rt5514 keeps power
but we shut down the i2c connection.

When this happens the rt5514 is confused about the state of the i2c
bus and won't recognize its own address.  That will lead to the rt5514
incorrectly NAKing the first transfer.

A single i2c transfer to any address should be enough to get the
rt5514 out of this funky state.

It is currently believed that this problem should be fixed in the
rt5514 driver itself because it seems that the i2c controller in the
rt5514 is easily confused.  Most i2c devices wouldn't detect a start
bit in this case.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7952b4baff402ddca1a263380bfd142f10290eb8)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>