[sound] alsa: asoc: rt298: fix capture doesn't work at some cases
Message-id: <1507829954160572-1077-patchset-jkysela@redhat.com>
Patchwork-id: 192669
O-Subject: [RHEL 7.5 PATCH v2 1076/1088] ALSA: ASoC: rt298: fix capture doesn't work at some cases
Bugzilla: 1463624
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RT298_CBJ_CTRL1(0x4f) bit 10 is needed for headset capture. It
will be turned off when "VREF" widget is on and be turned on when
bias level is ON. It is odd. And if "VREF" is turned on in bias
level is ON, RT298_CBJ_CTRL1(0x4f) bit 10 will be turned off.
This patch move the bit control from rt298_set_bias_level and
rt298_vref_event to rt298_jack_detect. So it will be turned on
once a jack is plugged in.

Signed-off-by: Bard Liao <bardliao@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 9ff49ce475cfeaf486321a2db8132a9500740faa)
Bugzilla: 1463624
Signed-off-by: Rafael Aquini <aquini@redhat.com>