[netdrv] igb: Use length to determine if descriptor is done
Message-id: <20171016152420.16656-9-vinschen@redhat.com>
Patchwork-id: 193698
O-Subject: [RHEL7.5 PATCH 08/37 v2] igb: Use length to determine if descriptor is done
Bugzilla: 1454902
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1454902
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14221296

commit 7ec0116c9131a8cd58dc456ae2bd5bc9976460d1
Author: Alexander Duyck <alexander.h.duyck@intel.com>

This change makes it so that we use the length of the packet instead of the
DD status bit to determine if a new descriptor is ready to be processed.
The obvious advantage is that it cuts down on reads as we don't really even
need the DD bit if going from a 0 to a non-zero value on size is enough to
inform us that the packet has been completed.

In addition I have updated the code so that we only reset the Rx descriptor
length for descriptor zero when resetting a ring instead of having to do a
memset with 0 over the entire ring.  By doing this we can save some time on
initialization.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

Signed-off-by: Corinna Vinschen <vinschen@redhat.com>
(cherry picked from commit 7ec0116c9131a8cd58dc456ae2bd5bc9976460d1)
Signed-off-by: Rafael Aquini <aquini@redhat.com>