[netdrv] igb: Add support for ethtool private flag to allow use of legacy Rx
Message-id: <20171016152420.16656-15-vinschen@redhat.com>
Patchwork-id: 193720
O-Subject: [RHEL7.5 PATCH 14/37 v2] igb: Add support for ethtool private flag to allow use of legacy Rx
Bugzilla: 1454902
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1454902
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14221296

commit e08912985b296b33b18a563cc126e3e2f018c2e1
Author: Alexander Duyck <alexander.h.duyck@intel.com>

Since there are potential drawbacks to the new Rx allocation approach I
thought it best to add a "chicken bit" so that we can turn the feature off
if in the event that a problem is found.

It also provides a means of validating the legacy Rx path in the event that
we are forced to fall back.  At some point in the future when we are
convinced we don't need it anymore we might be able to drop the legacy-rx
flag.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

Signed-off-by: Corinna Vinschen <vinschen@redhat.com>
(cherry picked from commit e08912985b296b33b18a563cc126e3e2f018c2e1)
Signed-off-by: Rafael Aquini <aquini@redhat.com>