[netdrv] i40e/i40evf: use cmpxchg64 when updating private flags in ethtool
Message-id: <20171017123911.21976-57-sassmann@redhat.com>
Patchwork-id: 194092
O-Subject: [RHEL7.5 PATCH 56/91] i40e/i40evf: use cmpxchg64 when updating private flags in ethtool
Bugzilla: 1454890
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

From: Jacob Keller <jacob.e.keller@intel.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1454890
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14279804

When a user gives an invalid command to change a private flag which is
not supported, either because it is read-only, or the device is not
capable of the feature, we simply ignore the request.

A naive solution would simply be to report error codes when one of the
flags was not supported. However, this causes problems because it makes
the operation not atomic. If a user requests multiple private flags
together at once we could end up changing one before failing at the
second flag.

We can do a bit better if we instead update a temporary copy of the
flags variable in the loop, and then copy it into place after. If we
aren't careful this has the pitfall of potentially silently overwriting
any changes caused by other threads.

Avoid this by using cmpxchg64 which will compare and swap the flags
variable only if it currently matched the old value. We'll report
-EAGAIN in the (hopefully rare!) case where the cmpxchg64 fails.

This ensures that we can properly report when flags are not supported in
an atomic fashion without the risk of overwriting other threads changes.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit 841c950d67c6facde32a8644ced20c04aebb7dd8)

Conflicts: Partial backport only.
Signed-off-by: Rafael Aquini <aquini@redhat.com>