[scsi] qla2xxx: Move #include qla_nvme.h to fix compile errors on RHEL 7
Message-id: <1508436256-8014-1-git-send-email-cdupuis@redhat.com>
Patchwork-id: 194586
O-Subject: [RHEL 7.5 e-stor V2 PATCH 22/23] qla2xxx: Move #include qla_nvme.h to fix compile errors on RHEL 7.
Bugzilla: 1316281
RH-Acked-by: Chris Leech <cleech@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

commit none

Move the #include of qla_nvme.h to after the system includes so that
qla_nvme.h has access to the scatter/gather list routines.

This patch is RHEL 7 specific as there are some block level changes to the NVMe
core which are not present in RHEL 7 including adding the defines for processing
scatter/gather lists.  It was agreed offline that the best to handle this is
by moving the inclusion of qla_nvme.h to after the scatterlist.h define so this
patch set only makes changes to the qla2xxx driver code itself.

Signed-off-by: Rafael Aquini <aquini@redhat.com>