[redhat] Integrate DWARF-based ABI comparison tool kabi-dw
Message-id: <20171020152845.27212-2-skozina@redhat.com>
Patchwork-id: 194629
O-Subject: [RHEL7.5 PATCH v4 1/1] Integrate DWARF-based ABI comparison tool kabi-dw
Bugzilla: 1499783
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Hannes Frederic Sowa <hannes@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

This patch integrates the kabi-dw tool into the kernel tree. There are two
intended ways to use the tool:

1) As a kernel developer running a local quick build you do:

$ make && make modules
$ make rh-kabi-dw-base  # To create the baseline kABI dataset
$ <apply your patch here>
$ make && make modules
$ make rh-kabi-dw-check # To generate a new dataset and compare it with the
                        # one stored previously

2) As a kernel maintainer preparing official build you generate the baseline
dataset for all arches by setting this macro in the kernel.spec file:

%define with_kabidw_base 1

Then you run a (scratch) brew build which will generate following sub-package
for all supported arches:

kernel-kabidw-base

The package contains the baseline dataset generated for all supported arches. If
the content of these subpackages is copied from /kabidw-base directory into the
redhat/kabi-dwarf/base directory of the kernel tree the following brew builds
will issue the comparison of the build with the baseline dataset. The report in
the brew build.log looks like this:

**** kABI DWARF-based comparison report ****
+ /builddir/build/BUILDROOT/kernel-3.10.0-681.el7.test.x86_64/kabi-dwarf/run_kabi-dw.sh compare /builddir/build/BUILDROOT/kernel-3.10.0-681.el7.test.x86_64/kabi-dwarf/base/x86_64 /builddir/build/BUILDROOT/kernel-3.10.0-681.el7.test.x86_64/kabi-dwarf/base/x86_64.tmp
Changes detected in: struct--pacct_struct.txt
Added:
+0x38 int this_is_newly_add_field_for_testing;
**** End of kABI DWARF-based comparison report ****

The kabi-dw tool is licensed under the GPL v3+ license except for the hash.[ch]
files which have been taken from libkmod and are licensed under the LGPL v2.1+.
Neither of this is a derived work of the Linux kernel.

The kabi-dw currently doesn't work best when multiple distinc definitions of
the structure are found within the kernel (eg. 'struct foo' is defined twice
with different layout). In such case one of these definitions is randomly
chosen for the comparison. Although this does happen in the Linux kernel (look
for 'struct shrinker' in 'include/drm/drm_backport.h' which leads to multiple
definitions of 'struct super_block') neither of such cases is currently
reachable from the whitelisted symbols and as such this issue does not impact
the report which is based on the kABI whitelisted symbols.

Signed-off-by: Rafael Aquini <aquini@redhat.com>