[scsi] scsi_dh: move .rescan to an auxiliary structure
Message-id: <1509120727-26247-56-git-send-email-snitzer@redhat.com>
Patchwork-id: 195295
O-Subject: [RHEL7.5 PATCH v2 55/58] scsi_dh: move .rescan to an auxiliary structure
Bugzilla: 1499107
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

All scsi_dh modules statically allocate a scsi_device_handler structure
so we need to add new members to an auxiliary structure that each device
handler allocates.  None of the scsi_dh interfaces are on the kABI
whitelist but we need to be defensive about a new RHEL7 kernel not
trying to call/access a scsi_device_handler structure member that
doesn't exist (e.g. the .rescan member would not exist in an old scsi_dh
module -- not that we know of any 3rd party scsi_dh modules but...).

An _optional_ scsi_device_handler_aux structure is passed to
scsi_register_device_handler() where it gets memcpy'd to a kzalloc'd
'struct scsi_device_handler_aux' and then gets added to a global
'scsi_dh_aux_list'.  The reason for the memcpy is to ensure that the
'struct scsi_device_handler_aux' associated with a scsi_dh module is
fully initialized (this avoids trying to access uninitialized memory).

scsi_rescan_device() is updated to access a scsi device's
scsi_device_handler_aux structure and if one exists it will call the
->rescan function.

scsi_unregister_device_handler() will take the scsi_device_handler_aux
off the 'scsi_dh_aux_list' and will free the memory allocated by
scsi_register_device_handler().

It should be noted that the above approach is an incremental improvement
to the model used in RHEL6.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>