[lib] idr: implement extended variant of idr
Message-id: <20171030162759.11090-1-ivecera@redhat.com>
Patchwork-id: 195379
O-Subject: [RHEL7.5 net PATCH v3 461/466] idr: implement extended variant of idr
Bugzilla: 1445420
RH-Acked-by: Stefano Brivio <sbrivio@redhat.com>

BZ#1445420
RHEL_only.

The patch introduces an extension to IDR that allows to store u32 indices
used by flower and net-sched actions (next patches in the series).
The current IDR implementation supports ID from range <0, INT_MAX> and
unfortunately classifiers and actions use u32's <0, UINT_MAX>. The IDR
in upstream was extended by commit 388f79fda74f ("idr: Add new APIs to
support unsigned long") that is not backportable currently because IDR
in upstream is based on radix-tree and RHEL uses an older implementation.
To backport this commit the IDR needs to be rebased and this is out of
scope of this series.
Instead of this I have created this extension that uses two IDR ranges
for requested range. The 1st is for <0, INT_MAX> and the 2nd is for
<INT_MAX+1, UINT_MAX>.

Differences between this and upstream:
The API in RHEL uses 'struct idr_ext' instead of 'struct idr'. To init
and destroy idr_init_ext() and idr_destroy_ext() are used instead of
idr_init() and idr_destroy(). RHEL support only range <0, UINT_MAX> that
is enough for net-sched.
The rest of API introduced by 388f79fda74f should be identical.

Changelog:
v2 (thx Stefano for review and comments):
* removed unnecessary explicit casts
* fixed multi-line comments
* added / replaced several comments per request
* fixed harmful bug in idr_remove_ext()
v3
* changed copyright notices

Cc: Stefano Brivio <sbrivio@redhat.com>
Cc: Jiri Benc <jbenc@redhat.com>

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>