[x86] intel_rdt/cqm: Add tasks file support
Message-id: <20171110101633.GD18088@krava>
Patchwork-id: 197520
O-Subject: [PATCHv2 RHEL7.5 16/42] x86/intel_rdt/cqm: Add tasks file support
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

upstream
========
commit d6aaba615a482ce7d3ec218cf7b8d02d0d5753b8
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Tue Jul 25 14:14:34 2017 -0700

RHEL7
=====
Adding new struct task_struct member 'rmid' as next KABI reserved item.

description
===========
The root directory, ctrl_mon and monitor groups are populated
with a read/write file named "tasks". When read, it shows all the task
IDs assigned to the resource group.

Tasks can be added to groups by writing the PID to the file. A task can
be present in one "ctrl_mon" group "and" one "monitor" group. IOW a
PID_x can be seen in a ctrl_mon group and a monitor group at the same
time. When a task is added to a ctrl_mon group, it is automatically
removed from the previous ctrl_mon group where it belonged. Similarly if
a task is moved to a monitor group it is removed from the previous
monitor group . Also since the monitor groups can only have subset of
tasks of parent ctrl_mon group, a task can be moved to a monitor group
only if its already present in the parent ctrl_mon group.

Task membership is indicated by a new field in the task_struct "u32
rmid" which holds the RMID for the task. RMID=0 is reserved for the
default root group where the tasks belong to at mount.

[tony: zero the rmid if rdtgroup was deleted when task was being moved]

Signed-off-by: Rafael Aquini <aquini@redhat.com>